implement reload; reorganize
