+++ app/src/main/res/layout/browser.xml
@@ -100,6 +100,7 @@
             android:layout_marginTop="5dp"
             android:layout_marginBottom="5dp"/>
         <TextView
+            android:id="@+id/reload"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/reload"
